+++ packages/perseus-core/src/index.ts
@@ -17,3 +17,5 @@ export {PerseusError} from "./error/perseus-error";
 export * from "./data-schema";
 
 export {pluck, mapObject, clone} from "./utils/object-utils";
+export {default as Util} from "./utils/util";
+export type {GridDimensions, Position} from "./utils/util";
