+++ packages/perseus/src/components/graphie-classes.ts
@@ -1,9 +1,7 @@
 /* eslint-disable @babel/no-invalid-this */
 
-import {Errors, PerseusError} from "@khanacademy/perseus-core";
+import {Errors, PerseusError, Util} from "@khanacademy/perseus-core";
 import _ from "underscore";
-
-import Util from "../util";
 
 const nestedMap = Util.nestedMap;
 const deepEq = Util.deepEq;
