+++ packages/perseus/src/components/graphie.tsx
@@ -1,11 +1,10 @@
-import {Errors} from "@khanacademy/perseus-core";
+import {Errors, Util} from "@khanacademy/perseus-core";
 import $ from "jquery";
 import * as React from "react";
 import _ from "underscore";
 
 import InteractiveUtil from "../interactive2/interactive-util";
 import {Log} from "../logging/log";
-import Util from "../util";
 import GraphUtils from "../util/graph-utils";
 import {Graphie as GraphieDrawingContext} from "../util/graphie";
