+++ packages/perseus/src/util/is-real-json-parse.ts
@@ -1,4 +1,4 @@
-import Util from "../util";
+import {Util} from "@khanacademy/perseus-core";
 
 const deepEq = Util.deepEq;
