+++ packages/perseus/src/widgets/cs-program/cs-program.tsx
@@ -2,6 +2,7 @@
  * This widget is for embedding Khan Academy CS programs.
  */
 
+import {Util} from "@khanacademy/perseus-core";
 import {StyleSheet, css} from "aphrodite";
 import $ from "jquery";
 import * as React from "react";
@@ -10,7 +11,6 @@ import _ from "underscore";
 import {getDependencies} from "../../dependencies";
 import * as Changeable from "../../mixins/changeable";
 import {articleMaxWidthInPx} from "../../styles/constants";
-import Util from "../../util";
 import {isFileProtocol} from "../../util/mobile-native-utils";
 import {toAbsoluteUrl} from "../../util/url-utils";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/cs-program/cs-program-ai-utils";
