+++ packages/perseus/src/widgets/interaction/interaction.tsx
@@ -2,12 +2,12 @@
 /* eslint-disable @babel/no-invalid-this, react/no-unsafe */
 import * as KAS from "@khanacademy/kas";
 import {vector as kvector} from "@khanacademy/kmath";
+import {Util} from "@khanacademy/perseus-core";
 import * as React from "react";
 import _ from "underscore";
 
 import Graphie from "../../components/graphie";
 import * as Changeable from "../../mixins/changeable";
-import Util from "../../util";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/interaction/interaction-ai-utils";
 import scoreNoop from "../__shared__/score-noop";
