+++ packages/perseus/src/widgets/interactive-graph.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable @babel/no-invalid-this, react/no-unsafe, react/sort-comp */
 import {number as knumber, point as kpoint} from "@khanacademy/kmath";
-import {Errors, PerseusError} from "@khanacademy/perseus-core";
+import {Errors, PerseusError, Util} from "@khanacademy/perseus-core";
 import $ from "jquery";
 import debounce from "lodash.debounce";
 import * as React from "react";
@@ -10,7 +10,6 @@ import Graph from "../components/graph";
 import {PerseusI18nContext} from "../components/i18n-context";
 import Interactive2 from "../interactive2";
 import WrappedLine from "../interactive2/wrapped-line";
-import Util from "../util";
 import KhanColors from "../util/colors";
 import {
     angleMeasures,
