+++ packages/perseus/src/widgets/plotter/validate-plotter.ts
@@ -1,4 +1,4 @@
-import Util from "../../util";
+import {Util} from "@khanacademy/perseus-core";
 
 import type {ValidationResult} from "../../types";
 import type {
